Fix compiler warnings across MADNESS codebase and applications
- Resolve -Wdeprecated-copy in Slice and CoeffTracker.
- Resolve -Wunused-parameter in stubmpi.h, gentensor.h, and function_interface.h.
- Resolve -Wrange-loop-construct in CalculationParameters.h and molecule.cc/h.
- Resolve -Wunused-variable / -Wunused-but-set-variable in operator.h, tensor_macros.h, ccpairfunction.cc, mp3.cc, GuessFactory.cc, test_jacobi.cc, test_tree_state.cc, response_state.hpp, fd_save_load.hpp, test_calc_manager.cpp, test_diagonalize_slots.cpp, Applications.hpp, and testpercoul.cc.
- Resolve -Wmaybe-uninitialized in worlddc.h and key.h.
- Resolve -Wsign-compare in vmra.h and test_memory_measurement.cc.
- Resolve -Wimplicit-fallthrough, -Wmisleading-indentation, and -Wcomment.

+++ src/madness/mra/vmra.h
@@ -574,7 +574,7 @@ namespace madness {
     	lindep *= s(s.size() - 1);  // eigenvalues are in ascending order
 
     	// transform s to s^{-1/2} in-place
-    	int rank = 0, nlindep = 0;
+    	size_t rank = 0, nlindep = 0;
     	for(size_t i = 0; i < n; ++i) {
             const auto s_i = s(i);
     		if (s_i > lindep) {
